requirements.txt
discord.py==1.4.1
PyNaCl==1.3.0
uvloop==0.14.0
python-dotenv==0.14.0
